+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 8.5.0.1 (Allied Bits)
-* Assignments should assign null if it is in the right operand.
+* Fixed the behavior of assignments which should assign null if it is in the right operand.
+* Fixed the comparison of an empty string with a numeric value: by default, an exception is thrown (as before), but if the "CompareIncompatibleTypes" flag is included in ExpressionOptions, comparison will evaluate to `false`.
 
 # 8.5.0 (Allied Bits)
 * Added support for user-defined functions.